Remove attributes spread for clarity
Initially spread {...attributes} for future
extensibility where I would merge commonProps with
common attributes for input. This is not being
implemented yet and is only causing confusion
and verbosity in the code. Remove for now until
future implementation.

Also extract initial form data and section heading
for clarity and readability.